+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   - added support to rotate Matrix4f
   - improved camera
   - added support to generate view matrix of camera with class TransformationUtils
+  - added support to set filter & wrap on texture
 
 ## Version 1.0.2 Alpha
   - added support to set an timer for an Runnable (execute Runnable after an fixe time)